requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.18.3
sqlalchemy>=1.4.47
psycopg2-binary>=2.9.6
python-dotenv>=0.21.0
pyspark>=3.3.2